implemented cli sq-addon-template

+++ hooks/post_gen_project.py
@@ -1,5 +1,5 @@
-#!/usr/bin/env python
-import os, shutil
+import os
+import shutil
 
 PROJECT_DIRECTORY = os.path.realpath(os.path.curdir)
 
@@ -10,7 +10,8 @@ def remove_dir(dir_path):
 
 
 def remove_file(filepath):
-    os.remove(os.path.join(PROJECT_DIRECTORY, filepath))
+    if os.path.isfile(filepath):
+        os.remove(os.path.join(PROJECT_DIRECTORY, filepath))
 
 
 if __name__ == '__main__':
@@ -18,10 +19,10 @@ if __name__ == '__main__':
     if '{{ cookiecutter.include_tests }}' not in ['yes', 'y']:
         remove_dir(os.path.join(PROJECT_DIRECTORY, 'tests'))
 
-    if '{{ cookiecutter.include_docs }}' not in ['yes', 'y']:
+    if '{{ cookiecutter.template_type }}' == 'add-on_code_only':
         remove_dir(os.path.join(PROJECT_DIRECTORY, 'docs_src'))
 
-    if '{{ cookiecutter.documentation_only }}' in ['yes', 'y']:
+    if '{{ cookiecutter.template_type }}' == 'documentation_only':
         remove_dir(os.path.join(PROJECT_DIRECTORY, 'seeq'))
         remove_dir(os.path.join(PROJECT_DIRECTORY, 'tests'))
         remove_dir(os.path.join(PROJECT_DIRECTORY, 'development'))
@@ -32,7 +33,7 @@ if __name__ == '__main__':
 
     if '{{ cookiecutter.project_license }}' == 'marketplace':
         doc_files_to_delete = ['github.md', 'docstrings.rst', 'add_on_installation.rst', 'backend_calculations.rst',
-                           'seeq_server_interactions.rst', 'user_interface.rst']
+                               'seeq_server_interactions.rst', 'user_interface.rst']
         for f in doc_files_to_delete:
             remove_file(os.path.join(PROJECT_DIRECTORY, 'docs_src', 'source', f))
